+++ backend/src/neuroagent/app/database/sql_schemas.py
@@ -14,7 +14,7 @@ from sqlalchemy import (
     Integer,
     String,
 )
-from sqlalchemy.dialects.postgresql import TSVECTOR
+from sqlalchemy.dialects.postgresql import JSONB, TSVECTOR
 from sqlalchemy.ext.asyncio import AsyncAttrs
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
@@ -94,7 +94,7 @@ class Messages(Base):
         DateTime(timezone=True), default=utc_now
     )
     entity: Mapped[Entity] = mapped_column(Enum(Entity), nullable=False)
-    content: Mapped[str] = mapped_column(String, nullable=False)
+    content: Mapped[JSONB] = mapped_column(JSONB, nullable=False)
     is_complete: Mapped[bool] = mapped_column(Boolean)
 
     thread_id: Mapped[uuid.UUID] = mapped_column(