+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Variable substitution in python tool.
+
 ## [v0.11.5] - 6.11.2025
 
 ### Changed
@@ -60,6 +63,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use typescript autogen for backend types in frontend.
 - Try to enforce using metric tools rather than downloading assets.
 - Rule to avoid overvalidating.
+- Plot through python + plotly.
+
+## Fixed
+- Images not appearing in Literature search (Prompt refactor).
+- Update uv lock package.
 
 ## [v0.10.0] - 2.10.2025